update models to use string values for totals, points, and prices

+++ model/points.go
@@ -1,5 +1,5 @@
 package model
 
 type PointsResponse struct {
-	Points float64 `json:"points"`
+	Points int `json:"points"`
 }
